初步完成秒杀逻辑，直接与mysql交互，未进行压力测试

+++ src/main/java/com/cai/seckill/pojo/User.java
@@ -2,14 +2,14 @@ package com.cai.seckill.pojo;
 
 public class User {
 
-    private Integer id;
+    private long id;
     private String name;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
